Move SILENT_LOGGER to execution environments package
