+++ packages/snaps-execution-environments/src/common/test-utils/logger.ts
