+++ packages/snaps-utils/src/test-utils/index.ts
@@ -5,7 +5,6 @@
  */
 
 export * from './common';
-export * from './logger';
 export * from './manifest';
 export * from './server';
 export * from './snap';
